requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
werkzeug==3.0.1
pandas>=2.2.0
openpyxl==3.1.2
reportlab==4.0.4
apscheduler==3.10.4
mercadopago>=2.2.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0